+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Check if etcd key `/config/_global/fqdn` exists before reading it; #21
 
 ## [v0.5.0](https://github.com/cloudogu/ces-commons/releases/tag/v0.5.0) - 2021-02-25
 ### Added